requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-genai==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
aiofiles==23.2.0
pydub==0.25.1
numpy==1.24.3
librosa==0.10.1
scikit-learn==1.3.0
python-dotenv==1.0.0
